Bad delaration.
git-svn-id: https://svn.forschooner.net/code/dnx/trunk/dnx@323 2be5450d-c085-40c6-83eb-9dc9c13f2dba

+++ server/dnxNebMain.c
@@ -1493,7 +1493,7 @@ unsigned long long dnxGetAffinity(char * name)
 // in the affinity bitmask.
 
 int dnxHammingWeight(unsigned long long x) {
-    uint64 count;
+    unsigned long long count;
     for (count=0; x; count++)
         x &= x-1;
     return count;
